requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
line-bot-sdk>=3.0.0
python-dotenv>=1.0.0
flask>=3.0.0
